requirements.txt
langchain
langchain-core
langchain-community
langgraph
sqlalchemy
langchain-postgres
langchain-mistralai
langchain-openai
psycopg2
huggingface
huggingface-hub
transformers
sqlalchemy-schemadisplay
graphviz==0.20.1
pydot==1.4.2
pyparsing==3.1.1
matplotlib
pandas
streamlit
tavily-python